Silenced alot of warnings on Clang (Mac)

+++ src/core/md_common.h
@@ -47,7 +47,7 @@
 #endif
 
 #ifndef ARRAY_SIZE
-#define ARRAY_SIZE(x) (sizeof(x)/sizeof(x[0]))
+#define ARRAY_SIZE(x) ((int64_t)(sizeof(x)/sizeof(x[0])))
 #endif
 
 #ifndef IS_ALIGNED
